requirements.txt
argon2-cffi==21.1.0
attrs==21.2.0
autopep8==1.6.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach==4.1.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
chart-studio==1.1.0
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1636043255643/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
defusedxml==0.7.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
Flask==2.0.2
Flask-Compress==1.10.1
fonttools==4.28.3
gif==3.0.0
graphviz==0.20
handcalcs==1.5.0
idna==3.3
innerscope==0.5.1
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1636998263526/work/dist/ipykernel-6.5.0-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1637978781393/work
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1637175074658/work
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1636814267019/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mistune==0.8.4
more-itertools==8.13.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
notebook==6.4.6
nptyping==1.4.4
numexpr==2.8.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
plotly==5.4.0
prometheus-client==0.12.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1636045889479/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycodestyle==2.8.0
pycparser==2.21
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1635877336018/work
requests==2.26.0
retrying==1.3.3
rich==10.14.0
scipy==1.7.3
Send2Trash==1.8.0
setuptools-scm==6.3.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1635819581833/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typish==1.9.3
urllib3==1.26.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
